refactor keyboard import
